fix(compile): resolve post-merge compilation errors
After merging #307-#316 in order, CI fails on compileDebugKotlin
with the following groups of errors:

1. MarkdownPreviewFragment.kt
   - Wrong import: androidx.compose.ui.text.util.Linkify
     → this package does not exist in compose; the actual
       Linkify.EMAIL_ADDRESSES / Linkify.WEB_URLS constants live in
       android.text.util.Linkify (Android framework).
   - Wrong import: com.itsaky.androidide.fragments.EditorFragmentTabManager
     → correct path is com.itsaky.androidide.fragments.editor.EditorFragmentTabManager
       (missing 'editor' sub-package).

2. BaseGitPageFragment.kt (addToolbarSectionLabel)
   - text = text infers TextView.text as val in this compose/kotlin
     combo, so the assignment is rejected. Use setText(text) explicitly.

3. resolveWorkspaceDirPath was duplicated in GitChangesFragment
   and GitBranchesFragment as private, and 2a2-A (#314) referenced
   it from GitHistoryFragment / GitStashFragment / GitDiffFragment
   where it wasn't visible. Promote it to a protected method on
   BaseGitPageFragment so every git page can use it.

4. GitBranchesFragment had a private emitGitOperation shadowing
   the one BaseGitPageFragment already exposes with the same
   signature. Delete the local copy (and the now-unused
   FragmentActivity/ViewModelProvider/IProjectManager imports).

5. GitChangesFragment: drop the now-orphaned IProjectManager import.

Net: -38 / +24, all 5 groups above go away.
